refactored get test stations to test station repository
