+++ tests/unit/CertificateGenerationService.unitTest.ts
@@ -425,79 +425,6 @@ describe('Certificate Generation Service', () => {
       });
     });
 
-    context('test getTestStation method', () => {
-      it('should return an array of test stations if invoke is successful', async () => {
-        const certGenSvc = Container.get(CertificateGenerationService);
-
-        const mockStations = testStationsMock;
-
-        invokeSpy
-          .mockResolvedValue({
-            Payload: toUint8Array(JSON.stringify({ body: JSON.stringify(mockStations) })),
-            FunctionError: undefined,
-            StatusCode: 200,
-          });
-
-        const testStations = await certGenSvc.getTestStations();
-
-        expect(testStations).toEqual(mockStations);
-        jest.clearAllMocks();
-      });
-      it('should invoke test stations up to 3 times if there is an issue', async () => {
-        const logSpy = jest.spyOn(console, 'error');
-        const certGenSvc = Container.get(CertificateGenerationService);
-
-        invokeSpy
-          .mockResolvedValue({
-            Payload: toUint8Array(JSON.stringify({ body: '' })),
-            FunctionError: undefined,
-            StatusCode: 200,
-          });
-
-        const testStations = await certGenSvc.getTestStations();
-
-        expect(logSpy).toHaveBeenLastCalledWith('There was an error retrieving the test stations on attempt 3: Error');
-        expect(logSpy).toHaveBeenCalledTimes(3);
-        expect(testStations).not.toBeNull();
-        logSpy.mockClear();
-        jest.clearAllMocks();
-      });
-      it('should return an empty array if test stations invoke is unsuccessful', async () => {
-        const certGenSvc = Container.get(CertificateGenerationService);
-
-        invokeSpy
-          .mockResolvedValue({
-            Payload: toUint8Array(JSON.stringify({ body: '' })),
-            FunctionError: undefined,
-            StatusCode: 200,
-          });
-
-        const testStations = await certGenSvc.getTestStations();
-
-        expect(testStations).toEqual([]);
-        jest.clearAllMocks();
-      });
-      it('should throw error if issue when parsing test stations', async () => {
-        const certGenSvc = Container.get(CertificateGenerationService);
-
-        const mockStations: ITestStation[] = [];
-
-        invokeSpy
-          .mockResolvedValue({
-            Payload: toUint8Array(JSON.stringify({ body: JSON.stringify(mockStations) })),
-            FunctionError: undefined,
-            StatusCode: 200,
-          });
-
-        const defects = await certGenSvc.getTestStations()
-          .catch((e) => {
-            expect(e).toBeInstanceOf(HTTPError);
-          });
-        expect(defects).toEqual(mockStations);
-        jest.clearAllMocks();
-      });
-    });
-
     context('test getDefectTranslations method', () => {
       it('should return an array of defects if invoke is successful', async () => {
         const certGenSvc = Container.get(CertificateGenerationService);
